Unused components was removed.

+++ base/src/main/java/com/resurrection/base/core/application/BaseApplication.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.os.Process
 import androidx.appcompat.app.AppCompatActivity
 import androidx.multidex.MultiDexApplication
+import com.resurrection.base.BuildConfig
 import com.resurrection.base.components.appstate.AppState
 import com.resurrection.base.components.crashtracker.CrashTrackerActivity
 import com.resurrection.base.components.dataholder.DataHolderManager
@@ -14,7 +15,7 @@ import com.resurrection.base.components.security.BiometricManager
 import com.resurrection.base.components.security.SecurityManager
 import com.resurrection.base.components.sharedpreferences.SharedPreferencesManager
 import com.resurrection.base.components.typeconverter.TypeConverter
-import com.resurrection.base.components.widget.AppLoadingIndicator
+import com.resurrection.base.components.widget.loadingindicator.LoadingIndicator
 import javax.inject.Inject
 import kotlin.system.exitProcess
 
@@ -31,9 +32,6 @@ abstract class BaseApplication : MultiDexApplication() {
 
     @Inject
     lateinit var loggerManager: LoggerManager
-
-    @Inject
-    lateinit var loadingIndicator: AppLoadingIndicator
 
     @Inject
     lateinit var networkManager: NetworkManager
@@ -59,7 +57,8 @@ abstract class BaseApplication : MultiDexApplication() {
     }
 
     open fun initCrashTracker(activityClass : Class<out AppCompatActivity> = CrashTrackerActivity::class.java){
-        Thread.setDefaultUncaughtExceptionHandler { t, e ->
+        Thread.setDefaultUncaughtExceptionHandler { thread, throwable ->
+            if (BuildConfig.DEBUG) throw throwable
             val intent = Intent(this, activityClass)
             this.startActivity(intent)
             Process.killProcess(Process.myPid())
